ai_interior_studio/requirements.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
Pillow
numpy
matplotlib
open3d
opencv-python
